backend/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.35.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
